app/requirements.txt
Flask==2.0.2
joblib==1.0.1
mlflow==1.23.1
numpy==1.19.5
PyYAML==6.0
scikit_learn==1.0.2
